Report IP of replying DNS-Server(s) in cases of hijack attemtps
Modify eventdns to let us know which DNS server answered the
request. Save the IP of the DNS-hijacking server to the log.
Seems to be the solution to bug2455, but at another glance
perhaps every DNS server should be logged?
Anyhow the 'struct nameserver' pointer is passed from eventdns
to tor code now and can be utilized appropriately everywhere.